+++ api/forms.py
@@ -2,12 +2,26 @@ from rest_framework import status
 from rest_framework.response import Response
 
 
-def handle_form(view, request):
+def handle_form(view, request, user=None):
+    """
+    Handle post requests in API by using django forms framework.
+
+    The form is populated with the posted data and any errors are returned as
+    JSON for the front end to display.
+    """
     try:
-        instance = view.model_class.objects.get(pk=request.data['pk'])
-        form = view.form_class(request.data, instance=instance)
+        # try and get an existing object, if it exists:
+        form_kwargs = {
+            'instance': view.model_class.objects.get(pk=request.data['pk']),
+        }
     except KeyError:
-        form = view.form_class(request.data)
+        # object, does not exist, we must be creating one:
+        form_kwargs = {}
+
+    if user is not None:
+        form_kwargs['user'] = user
+
+    form = view.form_class(request.data, **form_kwargs)
     if form.is_valid():
         form.full_clean()
         form.save()