+++ tests/functional_tests/utils.py
@@ -7,13 +7,21 @@ def click_and_wait(elem, t):
     return elem
 
 
+def get_closable_alerts(b):
+    alerts = b.find_elements_by_class_name('alert')
+    alerts = [alert for alert in alerts if alert.find_elements_by_class_name('fa-close')]
+    return alerts
+
+
 def check_and_close_msg(b, wait_time):
     """Check for presence of message and close."""
-    alert = b.find_elements_by_class_name('fa-close')
-    assert len(alert) == 1
-    click_and_wait(alert[0], wait_time)
-    alert = b.find_elements_by_class_name('fa-close')
-    assert len(alert) == 0
+    alerts = get_closable_alerts(b)
+    assert len(alerts) == 1
+    alert = alerts[0]
+    close_button = alert.find_elements_by_class_name('fa-close')
+    click_and_wait(close_button[0], wait_time)
+    alerts = get_closable_alerts(b)
+    assert len(alerts) == 0
 
 
 def assert_with_timeout(fn, max_t):