+++ CHANGELOG.md
@@ -3,6 +3,24 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [v2.7.0]
+
+### Added
+
+ - Add a new custom response for keywords: option to have a different reply to contacts that we have not seen before
+ - Add a `notes` field to contacts
+ - Add ansible support for deploying to Ubuntu 16.04
+ - Clear django cache on new ansible deploys
+
+### Changed
+
+ - Non-alphanumeric characters will be ignored at the start of a message when trying to match a keyword. E.g. `"keyword"` will now match `keyword`
+ - The contact edit page can now be viewed by any user with the `can_see_contact_names` permission - number and notes fields are only visible if the user has permission
+
+### Fixed
+
+ - Prevent users without correct permissions from changing phone numbers with the API
+
 ## [v2.6.3]
 
 ### Fixed